Add wheel build configuration to CI
This commit adds a new job to the release CI to leverage cibuildwheel to
build and publish portable wheels to PyPI. Right now at release time we
only publish an sdist which means users have to build from source when
then install mthree. By building wheels we ship precompiled binaries
that users can just install. This will build x86_64 and i686 on linux,
arm64 and x86_64 on MacOS, and amd64 and win32 for windows to start. We
can expand the supported architecture set later.