Only run wheel jobs in ci on release and publish to pypi
In #18 the PR was setup for interactive debugging and had removed the
sdist job, the upload to pypi and was also setup to run on every commit
so that I could debug issues. However, the intent of these jobs is to
only run at release for publishing wheels to pypi. This commit fixes the
definition so they only run on release and upload to pypi.