Improve numpy compatibility
This commit updates the build system to fix compatibility with numpy
versions. When we build wheels we should build against the oldest
version of numpy as possible. This ensures that when we publish wheels
to pypi the binaries will work with any version of numpy. Right now we
build against the latest version of numpy which restricts users to run
with a more recent version of numpy as it's built against a newer abi
version. To accomplish this a pyproject.toml is used, this makes it very
easy to list specific build system requirements and provides more
flexibility than calling pip during setup.py. This does require pip (or
any other install tool) to be compatible with PEP518 to be usable, but
that was initially included in pip 10, but it wasn't the best supported
until pip 19, which are both over 3 years old at this point and most
users building from source should hopefully have updated to at least
that version of pip.