DNM: Test wheel builds
This commit is not intended to be merged. It's sole purpose is to
manually trigger the wheel building jobs in CI prior to the upcoming
release. We recently made some changes to the ci configuration for the
wheel jobs and before we tag the release it would be good to test it.