Fix cibuildhweel configuration
In prepartion for the upcoming release we recently pushed #107 to test
the ci configuration for wheel jobs was fully working. That test run
showed a few small issues in the configuration which would have caused
an error at release time. Primarily adding a build skip for musl linux
environments, a test skip for py310 because of an orjson issue, and
changing the base image to manylinux2014 instead of manylinux2010. The
only one which will have user facing impact is the manylinux version
change. By switching to manylinux2014 this means the wheels are only
compatible with linux systems with the same (or newer) glibc as what
shipped in centos/rhel 7. So for users on older versions of linux
they'll have to build the latest release from source.