Only build release wheels with openmp on linux
As was discovered during the recent 0.24 release building with openmp
support on macOS and Windows fails to compile in the default
ci environment and the job will fail. This commit disables building with
openmp on those platforms so we only set the flag for linux builds. This
should fix future releases' builds so they build and publish without
issue.