Bump cibuildwheel version to latest
This commit bumps the cibuildwheel version to the latest release. This
enables building python 3.11 wheels in addition to just bumping the
python build toolchain we're using to build the wheels. Prior to merging
this commit we should do a manual test commit to run this on the PR and
not publish the build artifacts to ensure the job still works as
configured, and if there are any issues fix them on the branch prior to
reverting the test commit.