Add support for python 3.12 and fix build system
This commit makes two changes, the first is that it updates the package
metadata and ci configuration to add support for Python 3.12. On the
next release of mthree this will include a package for 3.12 users and
that we test 3.12 compatibility on PRs in CI.

Related to this a small fix was needed in the build system
configuration. The setuptools build-backend used for building mthree
with newer versions of setuptools was incompatible with the required
execution of the entire setup.py file (to generate the version.py
file). To fix this the build-backend is changed to use the legacy
setuptools backend which will work with the current setup for the
packaging. Longer term we should investigate updated the build
configuration to avoid needing to run the setup.py fully as part
of the package build.

Fixes #191