Fix release ci config issues, drop 3.7 support, and prepare 2.6.2 release
This commit fixes a release ci job configuration issue stemming from the
fact that the release ci jobs are setup to parallelize the sdist and
wheel publishing. The issue with this is the sdist, being just a
tarball, is significantly faster to publish and always goes first. But
the wheel builds which actually compile the cython code and run tests
are slower and much more fragile. We've had two releases in quick
succession where small oversights in the packaging configuration caused
the wheel jobs to fail, but by the time we hit the failure the sdist had
already been published to pypi which makes it hard to rollback. This
commit updates the ci configuration to fix this and only upload the
sdist after the wheels are there.

Then to prepare for the second patch release on 2.6.x this commit drops
support for python 3.7 and updates the ci jobs and package metadata to
indicate 3.7 is no longer supported and no longer build or test it.