Fix 'gmres' calls on old Scipy versions
Scipy 1.11 added the `rtol` option to `scipy.sparse.linalg.gmres`, which
is now the preferred form.  There are no prebuilt wheels for Scipy 1.11
for Python 3.8 on `pip` (and `mthree` lists its requirement as
`scipy>=1.3`), so this adds in a dynamic switch to the `gmres` function
to use the correct keyword argument based on the detect keyword
argument.